+++ src/commands/Themer.ts
@@ -1,5 +1,6 @@
 import * as vscode from 'vscode';
 import ChatClient from '../chat/ChatClient';
+import { join } from 'path';
 
 export class Themer {
 
@@ -16,7 +17,8 @@ export class Themer {
                                           f.packageJSON.contributes.themes.length > 0)
                              .forEach((fe: any) => {
                                 let iThemes = fe.packageJSON.contributes.themes.map((m: any) => {
-                                                                            return { id: fe.id, label: m.label };});
+                                                                            return { id: fe.id, label: m.label, themeId: m.id || m.label };});
+                                                                            
                                 this._availableThemes = (this._availableThemes.concat.apply(
                                                                                         this._availableThemes,
                                                                                         iThemes)
@@ -30,11 +32,15 @@ export class Themer {
         }
         param = param.toLowerCase().trim();
 
-        if (param === 'list') {
+        if (param === '') {
+            await this.currentTheme();
+        } else if (param === 'list') {
             // Whisper list of avaiable themes back to user
             await this.sendThemes(twitchUser);
         } else if (param === 'reset') {
             this.resetTheme();
+        } else if (param === 'random') {
+            this.randomTheme(twitchUser);
         } else {
             this.changeTheme(twitchUser, param);
         }
@@ -68,6 +74,13 @@ export class Themer {
         }
     }
 
+    private async randomTheme(twitchUser: string | undefined) {
+        const max = this._availableThemes.length;
+        const randomNumber = Math.floor(Math.random() * max);
+        const chosenTheme = this._availableThemes[randomNumber].label;
+        await this.changeTheme(twitchUser, chosenTheme);
+    }
+
     private async changeTheme(twitchUser: string | undefined, themeName: string) {
         // Find theme based on themeName and change theme if it is found
         let theme = this._availableThemes.filter(f => f.label.toLowerCase() === themeName.toLowerCase())[0];
@@ -77,8 +90,8 @@ export class Themer {
         
             if (x !== undefined) {
                 let conf = vscode.workspace.getConfiguration();
-                x.activate().then(f => {
-                    conf.update('workbench.colorTheme', theme.label, vscode.ConfigurationTarget.Global);
+                x.activate().then(f => { 
+                    conf.update('workbench.colorTheme', theme.themeId, vscode.ConfigurationTarget.Global);
                     if (twitchUser) {
                         vscode.window.showInformationMessage(`Theme changed to ${theme.label} by ${twitchUser}`);
                     }
@@ -86,11 +99,17 @@ export class Themer {
             }
         }
     }
+
+    private async currentTheme() {
+        const currentTheme = vscode.workspace.getConfiguration().get('workbench.colorTheme');
+        this._chatClient.sendMessage(`The current theme is ${currentTheme}`);
+    }
 }
 
 export interface ITheme {
     id: string;
     label: string;
+    themeId: string;
 }
 
 export interface IListRecipient {